ros_agent/requirements.txt
rclpy>=3.3.0
sensor_msgs>=4.2.0
nav_msgs>=4.2.0
geometry_msgs>=4.2.0
std_msgs>=4.2.0
requests>=2.28.0
pyyaml>=6.0
prompt_toolkit>=3.0.0
rich>=13.0.0
flask>=2.3.0
flask-socketio>=5.3.0

